Prepare file structure and get the flow ready to be run
